requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pymongo>=4.6.1
motor>=3.3.2
openai>=1.12.0
PyPDF2>=3.0.1
python-docx>=1.1.0
pydantic>=2.6.1
requests>=2.31.0
aiofiles>=23.2.1
mammoth>=1.6.0
fpdf2>=2.7.7
